+++ test/server.test.ts
@@ -1793,6 +1793,32 @@ describe("Dockerfile LSP Tests", function() {
 		lspProcess.on("message", semanticTokensListener);
 	});
 
+	it("issue #250", finished => {
+		this.timeout(5000);
+		sendNotification("textDocument/didOpen", {
+			textDocument: {
+				languageId: "dockerfile",
+				version: 1,
+				uri: "uri://dockerfile/250.txt",
+				text: "FROM scratch\nADD --chmod=644 . ."
+			}
+		});
+
+		lspProcess.once("message", json => {
+			if (json.method === "textDocument/publishDiagnostics") {
+				assert.equal(json.params.uri, "uri://dockerfile/250.txt");
+				console.log(json.params.diagnostics);
+				assert.equal(json.params.diagnostics.length, 0);
+				sendNotification("textDocument/didClose", {
+					textDocument: {
+						uri: "uri://dockerfile/250.txt"
+					}
+				});
+				finished();
+			}
+		});
+	});
+
 	function testInvalidFile(request: string, assertionCallback: Function) {
 		it(request, function(finished) {
 			this.timeout(5000);